+++ outputs.tf
@@ -13,7 +13,17 @@ output "postgres_instance_name" {
   value       = google_sql_database_instance.default.name
 }
 
+output "instance_name" {
+  description = "PSQL instance name"
+  value       = google_sql_database_instance.default.name
+}
+
 output "postgres_instance_connection_name" {
+  description = "PSQL instance connection name"
+  value       = google_sql_database_instance.default.connection_name
+}
+
+output "instance_connection_name" {
   description = "PSQL instance connection name"
   value       = google_sql_database_instance.default.connection_name
 }